idk wtf i'm doing...
I keep hearing that I'm not supposed to have too many elements (tsx) in one component, so I made 2 more.

Also, I learned how to use useContext. Yay! And I used it for everything. Absolutely everything...

I also had to make GlobalPosts.ts to specify the time of each variable / function. Not the kind of problem you would have in Java... Maybe not in Typescript either, but i might be too dumb for that.

Also, I am not sure in what component to put every function. Should there be so many in App.tsx?

Anyway, everything takes more space now and it's more complicated. I don't think this is really the conventional way, but I'll keep this branch for now.
ok bye